+++ SampleApp/SampleAppTests/TableViewControllerTests_XCTest.swift
@@ -7,12 +7,14 @@ import ExpectToEventuallyEqual
 import XCTest
 
 @MainActor
-final class TableViewControllerTests: XCTestCase, Sendable {
+final class TableViewControllerTests_XCTest: XCTestCase, Sendable {
     private var sut: TableViewController!
     private var tableDataSource: (any UITableViewDataSource)!
 
     override func setUp() async throws {
         try await super.setUp()
+        let book1 = SearchResult(trackName: "book 1", artistName: "author 1")
+        let book2 = SearchResult(trackName: "book 2", artistName: "Jon Reid")
         let fakeSearchProvider = FakeSearchProvider(searchResults: book1, book2)
         let storyboard = UIStoryboard(name: "Main", bundle: nil)
         sut = storyboard.instantiateViewController(identifier: String(describing: TableViewController.self))
@@ -43,20 +45,15 @@ final class TableViewControllerTests: XCTestCase, Sendable {
         )
     }
 
-    func test_secondRowShowsBookAuthor_FAILURE_DEMONSTRATION() async throws {
-        XCTExpectFailure("Demonstrate failure message")
+    func test_FAILURE_DEMONSTRATION_secondRowShowsBookAuthor() async throws {
         try await expectToEventuallyEqual(
             actual: { cellForRow(1).detailTextLabel?.text ?? "" },
             expected: "Steven Baker"
         )
     }
 
-    @MainActor
-    func cellForRow(_ row: Int) -> UITableViewCell {
+    private func cellForRow(_ row: Int) -> UITableViewCell {
         let indexPath = IndexPath(row: row, section: 0)
         return tableDataSource.tableView(sut.tableView, cellForRowAt: indexPath)
     }
 }
-
-let book1 = SearchResult(trackName: "book 1", artistName: "author 1")
-let book2 = SearchResult(trackName: "book 2", artistName: "Jon Reid")