+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+NEXT RELEASE
+------------
+
+- Breaking change: function now requires @MainActor
+- Added support for Swift Testing
+
+
 Version 1.1.0
 -------------
 _22 Dec 2023_